Time: 13 ms (91.64%), Space: 44.8 MB (29.31%) - LeetHub

+++ 0018-4sum/0018-4sum.java
@@ -22,7 +22,7 @@ class Solution {
             int l = start, r = nums.length-1;
             while(l<r){
                 if(nums[l] + nums[r] > target) r--;
-                else if ((long)(nums[l] + nums[r]) < target) l++;
+                else if ((nums[l] + nums[r]) < target) l++;
                 else{
                     List<Integer> temp = new ArrayList(list);
                     temp.add(nums[l]);
